convert `num_components` to `usize`

+++ garibaldi/opencv-rust/src/main.rs
@@ -3,8 +3,8 @@ use std::path::PathBuf;
 
 use opencv::{core, core::Vec3b, imgcodecs, imgproc, prelude::*, Result};
 
-fn random_color_map(num_labels: i32) -> Vec<Vec3b> {
-    let mut color_map: Vec<Vec3b> = vec![Vec3b::default(); num_labels as usize];
+fn random_color_map(num_labels: usize) -> Vec<Vec3b> {
+    let mut color_map: Vec<Vec3b> = vec![Vec3b::default(); num_labels];
     let mut rng = thread_rng();
     color_map[0] = Vec3b::default();
     for label in 1..num_labels {
@@ -19,7 +19,7 @@ fn random_color_map(num_labels: i32) -> Vec<Vec3b> {
     color_map
 }
 
-fn label_components(component_image: &Mat, num_components: i32) -> Result<Mat> {
+fn label_components(component_image: &Mat, num_components: usize) -> Result<Mat> {
     println!("num labels: {}", num_components);
 
     let color_map = random_color_map(num_components);
@@ -41,7 +41,7 @@ fn detect_components(src_path: PathBuf, components_path: PathBuf) -> Result<()> 
     imgproc::threshold(&src, &mut src_bw, 128.0, 255.0, imgproc::THRESH_BINARY)?;
     let mut component_image = unsafe { Mat::new_size(src.size()?, core::CV_32S)? };
     let num_components =
-        imgproc::connected_components(&src_bw, &mut component_image, 8, core::CV_32S)?;
+        imgproc::connected_components(&src_bw, &mut component_image, 8, core::CV_32S)? as usize;
 
     let label_image = label_components(&component_image, num_components)?;
